+++ Pilot2/P2B1/p2b1_default_model.txt
@@ -2,7 +2,7 @@
 num_hidden=[512,32]
 batch_size=32
 learning_rate=0.01
-epochs=2
+epochs=10
 cool=False
 weight_decay=0.01
 noise_factor=0.0
@@ -15,5 +15,3 @@ molecular_nbrs = 10
 molecular_epochs=10
 base_memo='p2b1'
 drop_prob = 0.5
-data_set='3k_Ordered'
-sampling_density = 0.1
